nove addReviewFunction to add-review server component
